Bug fixed
There was a tiny bug where the noise is twice the time of the original signal; that was because the convolution size became twice the original signal as the signal is convoluted with the response which is the same size of the original signal. This bug caused the output file to be twice the original file with half of it empty. Now this problem is solved.